# perf(rolldown_sourcemap): pre-allocate memory in `join` method