<!-- Thank you for contributing! -->

### Description

Local benchmark data

```
// Without pre-allocation
join/join               time:   [1.7542 ms 1.8115 ms 1.8710 ms]


// With pre-allocation
join/join               time:   [323.86 µs 324.37 µs 325.07 µs]
                        change: [−83.121% −82.664% −82.187%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
```

<!-- Please insert your description here and provide especially info about the "what" this PR is solving -->